+++ src/components/OptionMenu/OptionMenu.tsx
@@ -8,12 +8,21 @@ import { Portal } from 'react-portal';
 export interface IOption<V = any> {
   label: string;
   value: V;
-  onClick: (val: V) => void;
+  icon?: React.ComponentType;
+  onClick?: (val: V) => void;
+}
+
+export type TriggerProps = {
+  on?: boolean;
+  extraRound?: boolean;
+  buttonSize?: ControlSizes;
+  onClick?: React.MouseEventHandler<HTMLButtonElement>;
 }
 
 export type OptionMenuProps<V = any> = React.HTMLAttributes<HTMLDivElement> & {
   open?: boolean;
   optionMenuSize?: ControlSizes;
+  placement?: 'overlay' | 'below';
   stopPropagation?: boolean;
   portal?: HTMLElement | true;
   portalScrollParent?: HTMLElement;
@@ -25,6 +34,7 @@ export type OptionMenuProps<V = any> = React.HTMLAttributes<HTMLDivElement> & {
   extraRound?: boolean;
   options: IOption<V>[];
   optionListClassName?: string;
+  trigger?: (props: TriggerProps) => React.ReactNode;
   onOpen?: () => void;
   onClose?: () => void;
 }
@@ -41,8 +51,11 @@ export const OptionMenu = React.forwardRef<HTMLDivElement | null, OptionMenuProp
   portalScrollParent,
   portalScroll,
   hangLeft: hangLeftProp,
+  placement = 'overlay',
+  trigger,
   onOpen,
   onClose,
+  children,
   ...props
 }, ref) => {
   const OptionMenuRef = React.useRef<HTMLDivElement | null>(null);
@@ -76,6 +89,7 @@ export const OptionMenu = React.forwardRef<HTMLDivElement | null, OptionMenuProp
 
   const handleClick = React.useCallback((event: React.MouseEvent<HTMLButtonElement>) => {
     const rect = OptionMenuRef.current?.getBoundingClientRect();
+    console.log('rect', rect);
     if (stopPropagation) {
       event.stopPropagation();
       window.dispatchEvent(new Event('click'));
@@ -83,15 +97,15 @@ export const OptionMenu = React.forwardRef<HTMLDivElement | null, OptionMenuProp
     setIsOpen(!isOpenValue);
     if (rect) {
       setHangLeft(window.innerWidth < (rect.x + rect.width + 20));
-      setBoundingClientRet(portal ? rect : null);
+      setBoundingClientRet(rect);
     }
     if (isOpenValue && onClose) onClose();
     else if (!isOpenValue && onOpen) onOpen();
   }, [stopPropagation, isOpenValue, portal, boundingClientRect, onOpen, onClose]);
 
   const handleOptionClick = React.useCallback((opt: IOption) => {
-    opt.onClick(opt.value);
     setIsOpen(false);
+    if (opt.onClick) opt.onClick(opt.value);
     if (onClose) onClose();
   }, [onClose]);
 
@@ -100,6 +114,30 @@ export const OptionMenu = React.forwardRef<HTMLDivElement | null, OptionMenuProp
       setIsOpen(false);
     }
   }, [isOpenValue]);
+
+  const triggerElement = React.useMemo(() => {
+    if (trigger) {
+      return trigger({
+        on: isOpenValue,
+        extraRound: extraRound,
+        buttonSize: optionMenuSize,
+        onClick: handleClick,
+      })
+    }
+
+    return (
+      <IconButton
+        on={isOpenValue}
+        extraRound={extraRound}
+        buttonSize={optionMenuSize}
+        onClick={handleClick}
+      >
+        <svg width="15" height="15" viewBox="0 0 15 15" xmlns="http://www.w3.org/2000/svg">
+          <path fillRule="evenodd" clipRule="evenodd" d="M3 7.5C3 8.328 2.328 9 1.5 9C0.672 9 0 8.328 0 7.5C0 6.672 0.672 6 1.5 6C2.328 6 3 6.672 3 7.5ZM9 7.5C9 8.328 8.328 9 7.5 9C6.672 9 6 8.328 6 7.5C6 6.672 6.672 6 7.5 6C8.328 6 9 6.672 9 7.5ZM13.5 9C14.328 9 15 8.328 15 7.5C15 6.672 14.328 6 13.5 6C12.672 6 12 6.672 12 7.5C12 8.328 12.672 9 13.5 9Z" />
+        </svg>
+      </IconButton>
+    )
+  }, [trigger, isOpenValue, extraRound, optionMenuSize, handleClick]);
 
   const optionsList = React.useMemo(() => {
     const s_top = portalScroll
@@ -125,25 +163,28 @@ export const OptionMenu = React.forwardRef<HTMLDivElement | null, OptionMenuProp
             left: `${s_left + boundingClientRect.left}px`,
             minWidth: `${boundingClientRect.width}px`,
           } : {}),
+          ...(placement === 'below' ? {
+            transform: `translateY(${(boundingClientRect?.height ?? 0) + 4}px)`
+          } : {}),
         }}
       >
-        {options.map(o => (
+        {options.map(({ icon: Icon, ...opt }) => (
           <li
-            key={o.label}
+            key={opt.label}
             className={styles.option}
             onClick={(event: React.SyntheticEvent<HTMLLIElement>) => {
               if (stopPropagation) {
                 event.stopPropagation();
               }
-              handleOptionClick(o);
+              handleOptionClick(opt);
             }}
           >
-            {o.label}
+            {opt.label}
           </li>
         ))}
       </ul>
     );
-  }, [isOpenValue, stopPropagation, options, portal, hangLeftProp, portalScroll, portalScrollParent, optionMenuSize, optionListClassName, hangLeft, boundingClientRect, handleOptionClick]);
+  }, [isOpenValue, placement, stopPropagation, options, portal, hangLeftProp, portalScroll, portalScrollParent, optionMenuSize, optionListClassName, hangLeft, boundingClientRect, handleOptionClick]);
 
   React.useEffect(() => {
     window.addEventListener('resize', handleWindowResize);
@@ -173,16 +214,7 @@ export const OptionMenu = React.forwardRef<HTMLDivElement | null, OptionMenuProp
         !!extraRound && styles.extraRound,
       )}
     >
-      <IconButton
-        on={isOpenValue}
-        extraRound={extraRound}
-        buttonSize={optionMenuSize}
-        onClick={handleClick}
-      >
-        <svg width="15" height="15" viewBox="0 0 15 15" xmlns="http://www.w3.org/2000/svg">
-          <path fillRule="evenodd" clipRule="evenodd" d="M3 7.5C3 8.328 2.328 9 1.5 9C0.672 9 0 8.328 0 7.5C0 6.672 0.672 6 1.5 6C2.328 6 3 6.672 3 7.5ZM9 7.5C9 8.328 8.328 9 7.5 9C6.672 9 6 8.328 6 7.5C6 6.672 6.672 6 7.5 6C8.328 6 9 6.672 9 7.5ZM13.5 9C14.328 9 15 8.328 15 7.5C15 6.672 14.328 6 13.5 6C12.672 6 12 6.672 12 7.5C12 8.328 12.672 9 13.5 9Z" />
-        </svg>
-      </IconButton>
+      {triggerElement}
       {isOpenValue && (
         portal ? (
           <Portal node={portal === true ? document.body : portal}>