+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # CHANGELOG
 
+## v0.2.3
+* Fix `trigger` required in `OptionMenu` props
+
+## v0.2.2
+* Add `placement` option to `OptionMenu`
+* Added option to custom render `OptionMenu` trigger
+* Added `label` option for `IconButton`
+
 ## v0.2.1
 * Minor fix to `Button` props typing